twitter_be/requirement.txt
Flask==2.2.3
Flask_Admin==1.6.1
Flask_Cors==3.0.10
Flask_JWT_Extended==4.4.4
Flask_Login==0.6.2
Flask_Migrate==4.0.4
flask_sqlalchemy==3.0.5
minio==7.1.15
schedule==1.2.0
python-dotenv
gunicorn
